Display list of level one taxons
Display the list of level one taxons, taxons that have the GOV.UK homepage
as their parent taxon.

We had originally hoped to call content-store to get the content-item
for the homepage as this contains the list of level taxons tagged to the
new taxononmy, and then create a hard-coded list with the old style top level
taxons.

Unfortunately, we can't access the homepage content item until this PR:
https://github.com/alphagov/govuk-puppet/pull/7139 is merged and deployed.
Until then we are hard-coding all of the level one taxons.