Get the level one taxons from live content store
We have decided to get the published taxons from the live content 
store and one or two of the draft taxons from the draft content
store. This means that as more taxons are published we will have
more data to test in the prototype.

Removed the hard-coded list of root_taxons as we are no longer
using it.