Add tagged content list controller and change rummager search logic
This is to show a full tagged content items list by format.
The rummager search can filter by a number of field identifiers, so
we are filtering by format to return all the tagged content for a taxon
of a particular format.

Also made a change to the way we are stubbing call to rummager in tests
by using helper methods in the API Adaptors gem.

Add tagged content show view
Add class names using the BEM rules of CSS naming