Read level one taxons from multiple taxon files
Updated a test to check that we have the correct number of level one taxons. We are
assuming that each JSON file containing a branch of the site-wide taxonomy only ever
contains one level one taxon, so we can use the number of taxon files to determine if
we have the right number of level one draft taxons.