Use dynamic list of document formats
We don't want to hardcode a list of document formats into the prototype, as this would mean manually maintaining the list when formats are removed/added. We've chosen to use Rummager's list of 'migrated formats.yml' as this closely matches the formats that are being returned in our Rummager queries. However, there are some formats which are not listed in this yml file, so we are having to also hardcode those into our prototype